# api/openapi.yaml
openapi: 3.1.0
info:
  title: Vitamux API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >
    Contract-first API (ADR-0015, docs/architecture/api.md). /api/v1 is for the owner (session
    cookie + X-CSRF-Token, or an API key); /api/ingest/v1 is for clients and devices. Schemas
    marked as open objects get their properties when the endpoint is implemented; changes
    within v1 are additive only.
servers:
  - url: /
...
paths:
...
  /api/ingest/v1/devices/pair:
    post:
      tags: [ingest]
      operationId: pairDevice
      summary: Exchange a pairing code for a device token
      description: >
        Unauthenticated: the single-use code (POST /api/v1/devices/pairing-codes, valid for 10
        minutes) is the credential. Creates a device client on the owner's apple_health push
        connection, creating that connection the first time. Wrong codes are throttled per address
        (429 with Retry-After).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, name]
              properties:
                code: {type: string, minLength: 1, maxLength: 32, description: 'Case, spaces and dashes are ignored.'}
                name: {type: string, minLength: 1, maxLength: 100, description: Device name shown to the owner.}
      responses:
        '201':
          description: Paired. The token is shown only here.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DevicePairing'}
        '422': {$ref: '#/components/responses/Problem'}
        '429': {$ref: '#/components/responses/Problem'}
  /api/ingest/v1/devices/self:
    get:
      tags: [ingest]
      operationId: getDeviceSelf
      summary: The calling device's configuration, including pending anchor resets
      security:
        - clientToken: []
      responses:
        '200':
          description: Device configuration.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeviceSelf'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
...
  /api/ingest/v1/devices/self/rotate-token:
    post:
      tags: [ingest]
      operationId: rotateDeviceToken
      summary: Replace the calling device's token; the old one stops working at once
      security:
        - clientToken: []
      responses:
        '200':
          description: The new token, shown only here.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [token]
                properties:
                  token: {type: string, pattern: '^vmx_cli_[0-9a-f]{32}_[A-Za-z0-9_-]{43}$'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}

  # ---------------------------------------------------------------- catalogue
...
  /api/v1/devices:
    get:
      tags: [devices]
      operationId: listDevices
      summary: List paired devices, newest first, including revoked ones
      security: [{session: []}, {appSession: []}, {apiKey: ['read:config']}]
      responses:
        '200':
          description: Devices.
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices: {type: array, items: {$ref: '#/components/schemas/PairedDevice'}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
  /api/v1/devices/pairing-codes:
    post:
      tags: [devices]
      operationId: createPairingCode
      summary: Create a single-use pairing code, valid for 10 minutes
      description: At most 5 codes per 10 minutes (429 with Retry-After). 503 without VITAMUX_PUBLIC_URL, except for an app session, which redeems the code itself and gets no url or qr_payload then.
      security: [{session: [], csrf: []}, {appSession: []}, {apiKey: ['write:config']}]
      responses:
        '201':
          description: The code (shown once).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PairingCode'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '429': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /api/v1/devices/{id}/request-anchor-reset:
    post:
      tags: [devices]
      operationId: requestDeviceAnchorReset
      summary: Ask the device to resync types from scratch on its next contact
      description: >
        The device reads the request from GET /api/ingest/v1/devices/self. Re-sent samples are
        deduplicated by UUID.
      security: [{session: [], csrf: []}, {appSession: []}, {apiKey: ['write:config']}]
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                types:
                  type: array
                  maxItems: 64
                  description: HealthKit type identifiers; omitted or empty resets every type.
                  items: {type: string, pattern: '^HK[A-Za-z0-9]{1,126}$'}
      responses:
        '204':
          description: Requested.
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
...
  /api/v1/devices/{id}/revoke:
    post:
      tags: [devices]
      operationId: revokeDevice
      summary: Revoke a device's token; its next request is 401
      security: [{session: [], csrf: []}, {appSession: []}, {apiKey: ['write:config']}]
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Revoked.
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/origins:
    get:
      tags: [devices]
      operationId: listOrigins
      summary: List the apps (origins) data was recorded by, with their native or relayed state
      description: >
        Every origin seen inside a transport provider such as apple_health, by provider and key.
        relay_targets are the vendors an origin can be classified as relaying.
      security: [{session: []}, {appSession: []}, {apiKey: ['read:config']}]
      responses:
        '200':
          description: Origins and the vendors they may relay.
          content:
            application/json:
              schema:
                type: object
                required: [origins, relay_targets]
                properties:
                  origins: {type: array, items: {$ref: '#/components/schemas/DataOrigin'}}
                  relay_targets: {type: array, items: {$ref: '#/components/schemas/RelayTarget'}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
  /api/v1/origins/{id}:
    patch:
      tags: [devices]
      operationId: classifyOrigin
      summary: Set or clear the vendor an origin relays
      description: >
        The `relayed` rule selector and the all-sources view read the new state at once; resolved
        results are recomputed. The edit is audited and wins over the known relay defaults, which only seed
        new origins.
      security: [{session: [], csrf: []}, {appSession: []}, {apiKey: ['write:config']}]
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [relayed_provider]
              properties:
                relayed_provider:
                  type: [string, 'null']
                  description: A code from relay_targets, or null for an origin that records its own data.
      responses:
        '204':
          description: Classified.
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/source-devices:
    get:
      tags: [devices]
      operationId: listSourceDevices
      summary: List the devices measurements were recorded on (not the paired apps)
      description: >-
        The values the device_type, device_manufacturer and device_model rule selectors match (the rule builder's
        brand and device choices), the owner's names and merges, and the device types the owner can set. `include=records` adds each device's active records per connection.
      security: [{session: []}, {appSession: []}, {apiKey: ['read:config']}]
      parameters:
        - name: include
          in: query
          required: false
          description: Comma-separated expansions.
          style: form
          explode: false
          schema: {type: array, items: {type: string, enum: [records]}}
      responses:
        '200':
          description: Devices.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [devices, device_types]
                properties:
                  devices: {type: array, items: {$ref: '#/components/schemas/SourceDevice'}}
                  device_types:
                    type: array
                    items: {type: string}
                    description: The values PATCH accepts for device_type, which the built-in rules match.
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/source-devices/{id}:
    patch:
      tags: [devices]
      operationId: updateSourceDevice
      summary: Set a device's type or name (audited)
      description: >-
        Merge patch: a field present replaces the value, null clears it. A type set here wins over the one the
        normalizer reports (null hands it back); rules that select by device type are recomputed. 409 for a merged
        device: edit the device it was merged into.
      security: [{session: [], csrf: []}, {appSession: []}]
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SourceDevicePatch'}
      responses:
        '204':
          description: Updated.
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/source-devices/{id}/merge:
    post:
      tags: [devices]
      operationId: mergeSourceDevice
      summary: Merge a device into another device of the same provider (audited, irreversible)
      description: >-
        One transaction moves every record of the device (measurements, groups, sleep sessions, workouts, health
        events, superseded versions included) to the target, points the device and those already merged into it at
        the target, and marks the moved dates for recomputation. Later records of the device's fingerprint are written
        to the target. Dedupe keys keep the fingerprint, so reprocessing changes nothing. Rules that name the merged
        device by id stop matching it. 409 when either device is already merged; 422 for the device itself or a
        device of another provider.
      security: [{session: [], csrf: []}, {appSession: []}]
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [into]
              properties:
                into: {$ref: '#/components/schemas/DeviceID'}
      responses:
        '200':
          description: Merged; the rows moved per table.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [moved]
                properties:
                  moved: {$ref: '#/components/schemas/DeviceRecords'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}

  # ---------------------------------------------------------------- documents and labs
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: vitamux_session
      description: Owner session cookie set by POST /api/v1/auth/login (HttpOnly, Secure, SameSite=Strict).
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the session cookie on mutating requests; the value comes from login or GET /api/v1/auth/session.
    appSession:
      type: http
      scheme: bearer
      description: 'App session token `vmx_ses_<id>_<secret>` from POST /api/v1/auth/login with `client: app`. Admitted wherever the session is, without CSRF.'
    apiKey:
      type: http
      scheme: bearer
      description: 'API key `vmx_pat_<id>_<secret>`. The listed role is the scope it needs; `admin` implies every scope.'
    clientToken:
      type: http
      scheme: bearer
      description: 'Client token `vmx_cli_<id>_<secret>` with scope `ingest:<connection_id>`.'
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      schema: {type: string, minLength: 1}
...
  responses:
    Problem:
      description: Error (RFC 9457 problem+json; codes in internal/api/problem.go).
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        request_id: {type: string}
        errors:
          type: array
          items:
            type: object
            required: [pointer, detail]
            properties:
              pointer: {type: string}
              detail: {type: string}
...
    DeviceID:
      type: string
      pattern: '^dev_[0-9a-f]{32}$'
...
    PairedDevice:
      type: object
      description: A device paired with a pairing code (an ingest client of kind device).
      additionalProperties: false
      required: [id, name, connection_id, created_at, last_seen_at, last_sync_at, revoked_at, types, possibly_denied, anchor_resets]
      properties:
        id: {type: string, format: uuid, description: The device_id returned by pairing.}
        name: {type: string}
        connection_id: {type: string, pattern: '^conn_[0-9a-f]{32}$'}
        created_at: {type: string, format: date-time}
        last_seen_at: {type: [string, 'null'], format: date-time, description: Last authenticated request (to the minute).}
        last_sync_at: {type: [string, 'null'], format: date-time, description: When its newest batch arrived.}
        revoked_at: {type: [string, 'null'], format: date-time}
        types:
          type: array
          items: {type: string}
          description: HealthKit types in the connection's last healthkit.samples.v1 heartbeat checkpoint; empty until the device sends one.
        possibly_denied:
          type: array
          items: {type: string}
          description: >
            Requested types the connection stored nothing for in the last 7 days, while the device
            was paired for longer and seen within them. HealthKit does not reveal read denial, so
            this is a hint: a type can also be silent because there is nothing new to record. Types
            are tracked per connection, which paired devices share. Empty for a revoked device.
        anchor_resets: {type: array, items: {$ref: '#/components/schemas/AnchorReset'}}
    DataOrigin:
      type: object
      description: An app that recorded data inside a transport provider, e.g. a HealthKit bundle id.
      additionalProperties: false
      required: [id, provider, origin_key, name, is_native, relayed_provider, created_at]
      properties:
        id: {type: string, format: uuid}
        provider: {type: string, description: The transport provider code, e.g. apple_health.}
        origin_key: {type: string, description: 'The `origin_key` rule selector value.'}
        name: {type: [string, 'null']}
        is_native: {type: boolean, description: Recorded by the platform itself, such as Apple's own apps.}
        relayed_provider: {type: [string, 'null'], description: The vendor whose data the app relays; null when it records its own.}
        created_at: {type: string, format: date-time, description: When the origin was first seen.}
    RelayTarget:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: {type: string}
        name: {type: string}
    SourceDevice:
      type: object
      additionalProperties: false
      required: [id, provider, fingerprint, name, device_type, manufacturer, model, merged_into]
      properties:
        id: {$ref: '#/components/schemas/DeviceID'}
        provider: {type: string}
        fingerprint: {type: string, description: "The device's identity at the provider, as the normalizer reports it."}
        name: {type: [string, 'null'], description: The owner's label.}
        device_type: {type: [string, 'null'], description: 'The `device_type` rule selector value, e.g. watch.'}
        manufacturer: {type: [string, 'null'], description: 'The `device_manufacturer` rule selector value (the brand; matched case-insensitively).'}
        model: {type: [string, 'null'], description: 'The `device_model` rule selector value.'}
        merged_into:
          oneOf: [{$ref: '#/components/schemas/DeviceID'}, {type: 'null'}]
          description: The device this one was merged into; its records live there.
        connections:
          type: array
          description: 'With include=records: the connections with active records of the device, and how many.'
          items: {$ref: '#/components/schemas/DeviceConnectionRecords'}
    SourceDevicePatch:
      type: object
      description: Merge patch of a device; null clears a field.
      additionalProperties: false
      properties:
        device_type: {type: [string, 'null'], description: One of device_types from GET /source-devices.}
        name: {type: [string, 'null'], minLength: 1, maxLength: 100}
      x-go-type: json.RawMessage
    DeviceConnectionRecords:
      type: object
      additionalProperties: false
      required: [connection_id, records]
      properties:
        connection_id: {type: string, pattern: '^conn_[0-9a-f]{32}$'}
        records: {$ref: '#/components/schemas/DeviceRecords'}
    DeviceRecords:
      type: object
      additionalProperties: false
      description: >-
        Rows per table. Measurements count hourly aggregate rows plus daily values in a listing (an interval once per
        hour it touches, behind while aggregates rebuild), and rows in a merge.
      required: [measurements, groups, sleep_sessions, workouts, events]
      properties:
        measurements: {type: integer, format: int64}
        groups: {type: integer, format: int64}
        sleep_sessions: {type: integer, format: int64}
        workouts: {type: integer, format: int64}
        events: {type: integer, format: int64}
    AnchorReset:
      type: object
      additionalProperties: false
      required: [type, requested_at]
      properties:
        type: {type: string, description: 'A HealthKit type identifier, or * for every type.'}
        requested_at: {type: string, format: date-time, description: The latest request for this type.}
    PairingCode:
      type: object
      additionalProperties: false
      required: [code, expires_at]
      properties:
        code: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{4}-[0-9A-HJKMNP-TV-Z]{4}$', description: Crockford base32; single use.}
        expires_at: {type: string, format: date-time}
        url: {type: string, format: uri, description: The public base URL (VITAMUX_PUBLIC_URL) the app pairs against; omitted without one (app sessions only).}
        qr_payload: {type: string, description: 'Text for the QR code: the JSON object {"url", "code"} and nothing else; omitted with url.'}
    DevicePairing:
      type: object
      additionalProperties: false
      required: [device_id, connection_id, token]
      properties:
        device_id: {type: string, format: uuid}
        connection_id: {type: string, pattern: '^conn_[0-9a-f]{32}$'}
        token: {type: string, pattern: '^vmx_cli_[0-9a-f]{32}_[A-Za-z0-9_-]{43}$', description: 'Client token with scope ingest:<connection_id>; store it in the Keychain.'}
    DeviceSelf:
      type: object
      additionalProperties: false
      required: [device_id, connection_id, name, anchor_resets, source_filter]
      properties:
        device_id: {type: string, format: uuid}
        connection_id: {type: string, pattern: '^conn_[0-9a-f]{32}$'}
        name: {type: string}
        anchor_resets:
          type: array
          items: {$ref: '#/components/schemas/AnchorReset'}
          description: >
            Every reset the owner requested, kept by the server. Apply those whose requested_at is
            after the last reset applied for that type (* covers every type).
        source_filter: {$ref: '#/components/schemas/SourceFilter'}
    SourceFilterChoice:
      type: object
      description: The owner's explicit choice for one app in Apple Health.
      additionalProperties: false
      required: [bundle_id, mode]
      properties:
        bundle_id: {type: string, minLength: 1, maxLength: 255, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$', description: The HealthKit source's bundle id (the origin_key).}
        name: {type: string, maxLength: 200, description: The app's name as Apple Health shows it.}
        mode: {$ref: '#/components/schemas/SourceFilterMode'}
        types:
          type: array
          maxItems: 64
          description: 'per_type: the HealthKit types taken (the others are ignored). Dropped for take and ignore.'
          items: {type: string, pattern: '^HK[A-Za-z0-9]{1,126}$'}
...
    SourceFilter:
      type: object
      description: >
        What the device applies (docs/architecture/apple-health.md#source-filter). For an app,
        the first that matches decides: an explicit choice for its bundle id, then for its parent
        app (a Watch extension's bundle id minus .watchkitapp, .watchkitextension and similar);
        Apple's own sources (com.apple.*) are taken by default; a default_ignore pattern matching
        the bundle id or the parent's ignores it; anything else is taken.
      additionalProperties: false
      required: [version, origins, default_ignore]
      properties:
        version: {type: integer, minimum: 1, description: Incremented on every change of the explicit choices.}
        origins: {type: array, items: {$ref: '#/components/schemas/SourceFilterChoice'}}
        default_ignore: {type: array, items: {$ref: '#/components/schemas/SourceFilterDefault'}}
...